Fix sdk name and install on initialization
